+++ iam/policy/resource/s3/bucket/locals.tf
@@ -1,7 +1,11 @@
 locals {
-  bucket_arn   = startswith(var.bucket_name, "arn:") ? var.bucket_name : "arn:${var.std_map.iam_partition}:${local.service_name}:::${var.bucket_name}"
-  has_policy   = length(local.sid_list_single) != 0
-  policy_json  = local.has_policy ? jsonencode(module.this_policy["this"].iam_policy_doc) : data.aws_iam_policy_document.empty_policy["this"].json
+  bucket_arn = startswith(var.bucket_name, "arn:") ? var.bucket_name : "arn:${var.std_map.iam_partition}:${local.service_name}:::${var.bucket_name}"
+  has_policy = length(local.sid_list_single) != 0
+  policy_json = var.allow_access_point ? data.aws_iam_policy_document.ap_policy["this"].json : (
+    local.has_policy ? jsonencode(module.this_policy["this"].iam_policy_doc) : (
+      data.aws_iam_policy_document.empty_policy["this"].json
+    )
+  )
   service_name = "s3"
   sid_list_expanded = flatten([
     for v_sid in local.sid_list_single : [
@@ -26,16 +30,19 @@ locals {
       sid             = k_sid
     })
   ]
-  sid_list_single = !var.allow_public ? local.sid_list_single_var : concat(local.sid_list_single_var, [
-    {
-      access          = "public_read"
-      condition_map   = {}
-      identifier_list = ["*"]
-      identifier_type = "*"
-      object_key_list = ["*"]
-      sid             = "World"
-    }
-  ])
+  sid_list_single = concat(
+    local.sid_list_single_var,
+    var.allow_public ? [
+      {
+        access          = "public_read"
+        condition_map   = {}
+        identifier_list = ["*"]
+        identifier_type = "*"
+        object_key_list = ["*"]
+        sid             = "World"
+      }
+    ] : [],
+  )
   sid_map = {
     for v_sid in local.sid_list_expanded : v_sid.sid => v_sid
   }